batchfetch export .result
